// src/lib/generated/models/transactionResponseDtoType.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API Finanças Pessoais
 * API de controle de finanças pessoais com categorias, transações, orçamentos e relatórios
 * OpenAPI spec version: 1.0
 */

/**
 * Tipo de transação
 */
export type TransactionResponseDtoType = typeof TransactionResponseDtoType[keyof typeof TransactionResponseDtoType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const TransactionResponseDtoType = {
  INCOME: 'INCOME',
  EXPENSE: 'EXPENSE',
} as const;